refactor: update path config
