+++ apps/user/src/app.module.ts
@@ -1,11 +1,10 @@
+import { DatabaseModule } from '@libs/common-db/database.module';
 import { Module } from '@nestjs/common';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { MongooseModule } from '@nestjs/mongoose';
+import { ConfigModule } from '@nestjs/config';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
-import { DatabaseModule } from './database/database.module';
 import { TaskModule } from './task/task.module';
 import { UserModule } from './user/user.module';
 import { VpnModule } from './vpn/vpn.module';
@@ -13,12 +12,6 @@ import { VpnModule } from './vpn/vpn.module';
 @Module({
   imports: [
     ConfigModule.forRoot({ isGlobal: true }),
-    MongooseModule.forRootAsync({
-      inject: [ConfigService],
-      useFactory: (config: ConfigService) => ({
-        uri: config.get('USER_DB_URI'),
-      }),
-    }),
     UserModule,
     AuthModule,
     DatabaseModule,
